update redux branch
